moving join game to folder
